// frontend/src/glorp/types.ts
// <auto-generated>
// Generated by Glorp.Api TypeScriptGenerator on 27/05/2026 09:38:11.
// Do not edit by hand. Regenerate: run api project with "generate-client".
// </auto-generated>

/**
 * A Bar
 */
export interface Bar {
  name: string;
  height: number;
  weight: number;
}

/**
 * A Glorp Bars request object, returns any Bar with a height greater than or equal to the provided MinHeight
 */
export interface BarsRequest {
  $type: "BarsRequest";
  minHeight: number;
}

/**
 * A Glorp Bars response object
 */
export interface BarsResponse {
  $type: "BarsResponse";
  success: boolean;
  message?: string | null;
  data?: Bar[] | null;
}

/**
 * A Foo
 */
export interface Foo {
  name: string;
  color: string;
}

/**
 * A Glorp Foos request object, returns any Foo with a name that contains the provided Name string (case-insensitive)
 */
export interface FoosRequest {
  $type: "FoosRequest";
  name: string;
}

/**
 * A Glorp Foos response object
 */
export interface FoosResponse {
  $type: "FoosResponse";
  success: boolean;
  message?: string | null;
  data?: Foo[] | null;
}

export type GlorpRequest = BarsRequest | FoosRequest;
export type GlorpResponse = BarsResponse | FoosResponse;

export interface GlorpRequestResponseMap {
  BarsRequest: BarsResponse;
  FoosRequest: FoosResponse;
}

export type GlorpResponseFor<TReq extends GlorpRequest> =
  TReq["$type"] extends keyof GlorpRequestResponseMap
    ? GlorpRequestResponseMap[TReq["$type"]]
    : never;